Provide VB implementation of 'use auto property'
