+++ src/EditorFeatures/VisualBasic/BasicEditorFeatures.vbproj
@@ -241,6 +241,9 @@
     <Compile Include="SignatureHelp\SignatureHelpUtilities.vb" />
     <Compile Include="TextStructureNavigation\TextStructureNavigatorProvider.vb" />
     <Compile Include="TodoComment\BasicTodoCommentIncrementalAnalyzerProvider.vb" />
+    <Compile Include="UseAutoProperty\UseAutoPropertyAnalyzer.vb" />
+    <Compile Include="UseAutoProperty\UseAutoPropertyCodeFixProvider.vb" />
+    <Compile Include="UseAutoProperty\Utilities.vb" />
     <Compile Include="Utilities\LineAdjustmentFormattingRule.vb" />
     <Compile Include="Utilities\CommandHandlers\AbstractImplementAbstractClassOrInterfaceCommandHandler.vb" />
     <Compile Include="Utilities\NamedTypeSymbolExtensions.vb" />
